QUERIES UPDATED

+++ sqlParser/test_queries.sql
@@ -44,17 +44,16 @@ FHC		varchar(3),
 LAND		varchar(3),
 STADT		varchar(50) ,
 NAME		varchar(50) ,
-constraint FLUGHAFEN_PS
-		primary key (FHC),
+constraint FLUGHAFEN_PS primary key (FHC),
 );
 
 create table FLUG (
-FNR             integer,
+FNR    integer,
 FLC		varchar(2),
 FLNR		integer,
 VON		varchar(3),
 NACH		varchar(3),
-AB		integer,
+AB		integer(77),
 AN		integer,
 constraint FLUG_PS
 		primary key (FNR),
